refactor: move PersonalDetails store to hooks.ts file
